qlora_finetune/requirements.txt
# Core dependencies from main project
# Note: Install PyTorch from CUDA 12.1 index for compatibility
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio --index-url https://download.pytorch.org/whl/cu121
torch==2.1.0
torchvision==0.16.0
torchaudio
transformers==4.34.1  # Match main project version
tokenizers>=0.14,<0.15  # Required by transformers 4.34.1
accelerate==0.26.1  # Compatible with transformers 4.34.1
huggingface_hub==0.19.4  # Compatible with transformers 4.34.1 and datasets 2.17.1
numpy>=1.17,<2.0.0  # Required by accelerate
dill<0.3.9,>=0.3.0  # Required by datasets 2.17.1
fsspec>=2023.1.0,<2023.11.0  # Required by datasets 2.17.1
triton==2.1.0  # Required by torch 2.1.0
einops
timm==0.9.10
pillow
wandb

# QLoRA specific dependencies
peft==0.7.1  # Compatible with transformers 4.34.1
bitsandbytes>=0.41.0
trl==0.7.4  # Compatible with transformers 4.34.1
datasets==2.17.1  # Compatible with huggingface_hub 0.19.4

# Additional utilities
jsonlines
rich
tqdm

# Mamba SSM dependencies (required for Cobra models)
mamba-ssm<2.0.0
causal-conv1d

